requirements.txt
streamlit==1.38.0
tensorflow-cpu==2.20.0
numpy==2.1.0
pillow==10.0.1